+++ client/App.tsx
@@ -8,14 +8,15 @@ import { SSRProvider } from "./hooks/useSSR";
 import { ConfigContext } from "./hooks/useConfig";
 import { CSRFContext } from "./hooks/useCSRF";
 import { ThemeProvider } from "./hooks/useTheme";
+import { PostsCacheProvider } from "./hooks/usePostsCache";
 import requestJSON from "./helpers/requestJSON";
 import TestPage from "./routes/TestPage";
 import IndexPage from "./routes/index/IndexPage";
 import SearchPage from "./routes/search/SearchPage";
 import PostPage from "./routes/post/PostPage";
 import TagsPage from "./routes/tags/TagsPage";
-import "./globals.scss";
 import NotFoundPage from "./routes/error/NotFoundPage";
+import "./globals.scss";
 
 interface Props {
   initialData: AnySSRPageData;
@@ -70,15 +71,17 @@ export default hot(module)(function App({ initialData }: Props) {
         <CSRFContext.Provider value={initialData._csrf}>
           <ConfigContext.Provider value={initialData._config}>
             <PageDataContext.Provider value={contextData}>
-              <Switch>
-                <Route path="/test" component={TestPage} />
-                <Route path="/tags" component={TagsPage} />
-                <Route path="/posts/:id" component={PostPage} />
-                <Route path="/posts" component={SearchPage} />
-                <Route path="/" exact component={IndexPage} />
-                <NotFoundPage />
-              </Switch>
-              <ToastContainer />
+              <PostsCacheProvider>
+                <Switch>
+                  <Route path="/test" component={TestPage} />
+                  <Route path="/tags" component={TagsPage} />
+                  <Route path="/posts/:id" component={PostPage} />
+                  <Route path="/posts" component={SearchPage} />
+                  <Route path="/" exact component={IndexPage} />
+                  <NotFoundPage />
+                </Switch>
+                <ToastContainer />
+              </PostsCacheProvider>
             </PageDataContext.Provider>
           </ConfigContext.Provider>
         </CSRFContext.Provider>